requirements.txt
fastapi==0.104.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
passlib==1.7.4
bcrypt==4.3.0
python-multipart==0.0.20
pydantic==2.11.7
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1